runner implementations
This adds a runner class that is responsible for running the script.
Running involves running each phase (dump, copy, restore) and firing
callbacks along the way.  The runner also runs a setup and teardown
that ensures the output dirs for both the source and target are available
when needed and cleaned up when running is complete.