updated to the latest ns-options
This version fixes some issues where inherited options don't get the
definition from the superclass definition.  This was breaking some
stuff and causing me to have to do some dirty hacks.

With the move to this version, I can remove the hacks and what-not.

This also upgrades to the latest scmd.