Hotfix: Update to the latest scmd
This updates to the latest scmd. This ensures dumpdb works with
it and allows other repos to update to the latest scmd that are
also using dumpdb.

This also locks down pry and rake to avoid newer versions that
require a new ruby version and this matches our current
conventions.

/cc @kellyredding