Remove configuring yaml databases
This removes configuring yaml databases and being able to use a
"database name" in the source and target. This is to simplify
dumpdb and avoid all the yaml parsing logic. A `database.yml` can
still be used with dumpdb it just has to be manually parsed and
used in the source and target.

The source and target now always expect a hash of values. This
works as it previously did. A `Db` object is built and provided as
the source and target in an initialized dumpdb script.

Additionally, the source and target can now be configured with a
port. This previously wasn't an option but it now is.

This also includes a few internal cleanups to the `Db` class and
updates its tests to our latest conventions. It now uses symbol
keys instead of string keys which is just to match how it hsubs its
values into the cmd templates. It also no longer supports the
aliases for its values that it previously did. The other tests
could be updated to our latest conventions but we don't want to
worry about making those changes right now.