(hotfix) update/cleanup tests; get up to our latest standards
The tests here were pretty dated.  There were many style issues
and organizational choices that weren't up to our latest standards
and conventions.  This does a pass through the code to clean
things up a bit.

This is prep for removing ns-options in a future effort.  I want
to do this first so this noise won't junk up that future PR.