remove ns-options as a dependency
This is part of getting this gem working in modern ruby versions.
We are moving away from using ns-options as a dependency b/c it
has performance issues when accessing options and on top of that
its API hasn't proven to be especially useful over a struct-like
class.

This removes the `:settings` options group.  Since it was just
handling option storage, I replaced it with a basic hash with
the appropriate defaults set for each option value.

I also brought in much-plugin since this is a mixin with an
included hook.